refactor(checkbox): unify platform-specific files (use Android impl)
Per maintainer feedback on #4955, drop the iOS-specific Checkbox renderer
and inline the unified MD3 implementation directly into Checkbox.tsx.
The same MaterialCommunityIcon-based control now renders on both
platforms.

Changes:
- Checkbox.tsx: inlines what was the Android implementation; renders the
  same control on both platforms. Drops the Platform.OS dispatching.
- CheckboxAndroid.tsx, CheckboxIOS.tsx: deleted.
- Checkbox/index.ts: `Checkbox.Android` and `Checkbox.IOS` are kept as
  back-compat aliases of `Checkbox` itself, so existing imports keep
  working with no breaking change.
- CheckboxItem.tsx: drops the `mode`-based branching; always renders
  `<Checkbox />`. The `mode` prop is kept and marked deprecated.
- src/index.tsx: `CheckboxAndroidProps` and `CheckboxIOSProps` now
  alias the unified `Props` from Checkbox.tsx (no breaking change).
- docs/docusaurus.config.js: removed entries for the now-deleted
  CheckboxAndroid and CheckboxIOS files.
- Snapshot tests updated: the renderer now uses the unified container
  (36x36 ripple) and MaterialCommunityIcon glyphs on both platforms.

`getSelectionControlIOSColor` in utils.ts is intentionally left in
place because RadioButtonIOS still depends on it. A similar unification
for RadioButton can follow in a separate PR.

As a side benefit, callers using `<Checkbox.IOS uncheckedColor=...>`
will now see the prop applied (the old iOS variant silently dropped it),
addressing one of the API-parity gaps noted in #4949.
