--- src/components/Checkbox/CheckboxIOS.tsx
@@ -1,116 +1,0 @@
-import * as React from 'react';
-import {
-  ColorValue,
-  GestureResponderEvent,
-  StyleSheet,
-  View,
-} from 'react-native';
-
-import { getSelectionControlIOSColor } from './utils';
-import { useInternalTheme } from '../../core/theming';
-import type { $RemoveChildren, ThemeProp } from '../../types';
-import MaterialCommunityIcon from '../MaterialCommunityIcon';
-import TouchableRipple from '../TouchableRipple/TouchableRipple';
-
-export type Props = $RemoveChildren<typeof TouchableRipple> & {
-  /**
-   * Status of checkbox.
-   */
-  status: 'checked' | 'unchecked' | 'indeterminate';
-  /**
-   * Whether checkbox is disabled.
-   */
-  disabled?: boolean;
-  /**
-   * Function to execute on press.
-   */
-  onPress?: (e: GestureResponderEvent) => void;
-  /**
-   * Custom color for checkbox.
-   */
-  color?: ColorValue;
-  /**
-   * Whether the checkbox is in an error state. When true, the checked /
-   * indeterminate icon uses `theme.colors.error`. `disabled` and explicit
-   * `color` overrides take precedence.
-   */
-  error?: boolean;
-  /**
-   * @optional
-   */
-  theme?: ThemeProp;
-  /**
-   * testID to be used on tests.
-   */
-  testID?: string;
-};
-
-/**
- * Checkboxes allow the selection of multiple options from a set.
- * This component follows platform guidelines for iOS, but can be used
- * on any platform.
- *
- * @extends TouchableRipple props https://callstack.github.io/react-native-paper/docs/components/TouchableRipple
- */
-const CheckboxIOS = ({
-  status,
-  disabled,
-  onPress,
-  theme: themeOverrides,
-  testID,
-  error,
-  ...rest
-}: Props) => {
-  const theme = useInternalTheme(themeOverrides);
-  const checked = status === 'checked';
-  const indeterminate = status === 'indeterminate';
-
-  const { checkedColor } = getSelectionControlIOSColor({
-    theme,
-    disabled,
-    customColor: rest.color,
-    error,
-  });
-
-  const icon = indeterminate ? 'minus' : 'check';
-  const opacity = indeterminate || checked ? 1 : 0;
-
-  return (
-    <TouchableRipple
-      {...rest}
-      borderless
-      onPress={onPress}
-      disabled={disabled}
-      accessibilityRole="checkbox"
-      accessibilityState={{ disabled, checked }}
-      accessibilityLiveRegion="polite"
-      style={styles.container}
-      testID={testID}
-      theme={theme}
-    >
-      <View style={{ opacity }}>
-        <MaterialCommunityIcon
-          allowFontScaling={false}
-          name={icon}
-          size={24}
-          color={checkedColor}
-          direction="ltr"
-        />
-      </View>
-    </TouchableRipple>
-  );
-};
-
-CheckboxIOS.displayName = 'Checkbox.IOS';
-
-const styles = StyleSheet.create({
-  container: {
-    borderRadius: 18,
-    padding: 6,
-  },
-});
-
-export default CheckboxIOS;
-
-// @component-docs ignore-next-line
-export { CheckboxIOS };
