+++ src/components/Checkbox/index.ts
@@ -1,6 +1,4 @@
 import CheckboxComponent from './Checkbox';
-import CheckboxAndroid from './CheckboxAndroid';
-import CheckboxIOS from './CheckboxIOS';
 import CheckboxItem from './CheckboxItem';
 
 const Checkbox = Object.assign(
@@ -9,10 +7,6 @@ const Checkbox = Object.assign(
   {
     // @component ./CheckboxItem.tsx
     Item: CheckboxItem,
-    // @component ./CheckboxAndroid.tsx
-    Android: CheckboxAndroid,
-    // @component ./CheckboxIOS.tsx
-    IOS: CheckboxIOS,
   }
 );
