+++ src/components/__tests__/Checkbox/CheckboxItem.test.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { Platform } from 'react-native';
 
 import { act, fireEvent } from '@testing-library/react-native';
 
@@ -14,31 +13,11 @@ it('renders unchecked', () => {
   expect(tree).toMatchSnapshot();
 });
 
-it('can render the iOS checkbox on different platforms', () => {
-  Platform.OS = 'android';
-  const tree = render(
-    <Checkbox.Item status="unchecked" label="iOS Checkbox" mode="ios" />
-  ).toJSON();
-
-  expect(tree).toMatchSnapshot();
-});
-
-it('can render the Android checkbox on different platforms', () => {
-  Platform.OS = 'ios';
-  const tree = render(
-    <Checkbox.Item status="unchecked" label="iOS Checkbox" mode="android" />
-  ).toJSON();
-
-  expect(tree).toMatchSnapshot();
-});
-
 it('can render leading checkbox control', () => {
-  Platform.OS = 'ios';
   const tree = render(
     <Checkbox.Item
       label="Default with leading control"
       status={'unchecked'}
-      mode="ios"
       position="leading"
     />
   ).toJSON();
